zio_compress: introduce compress_threshold
Now default compression is lz4, which can stop
compression process by itself on incompressible data.

We already have 12.5% compress threshold, and we can't
just remove it because there'll be worst cases with
additional CPU usage on decompression of previously
uncompressed data, so leave it for future.

But there are two cases, which we can optimize:
- less than BPE_PAYLOAD_SIZE (embedded_data feature);
- at least one saved sector (don't waste CPU time).

If data wasn't compressed - it will be saved as
ZIO_COMPRESS_OFF, so if we really need to recompress
data without ashift info and check anything -
we can just compress it with zero threshold.
So, we don't need a new feature flag here!

Signed-off-by: George Melikov <mail@gmelikov.ru>