zio_compress: use spa_max_ashift as threshold
Now default compression is lz4, which can stop
compression process by itself on incompressible data.
If there is additional size checks -
we will only make our compressratio worse.
So we can check only for sector size.

If data wasn't compressed - it will be saved as
ZIO_COMPRESS_OFF, so if we really need to recompress
data without ashift info and check anything -
we can just compress it with zero threshold.
So, we don't need a new feature flag here!

Signed-off-by: George Melikov <mail@gmelikov.ru>